moving all logic to app
